Add initial date to options
Added initial date option to allow you to set an initial date to display
in the calendar that is not the current date.  Also added a method to be
able to update that initial date.  This is very useful when you are
within another view that uses dates to display information and you want
to display the datepicker with an initial date on that view.